fix: connection leak during listener close
Remove Close from LimitListener, falling back on the implementation by the
underlying net.Listener. The done channel was needed in the original
LimitListener due to it blocking before calling the underlying Listener's
Accept. Because closing the underlying listener will cause our call to its
Accept to return with an error that we pass along, we don't need to care at
all about it.

Signed-off-by: Erik Swanson <erik@retailnext.net>